requirements.txt
Flask
Flask-Session
Flask-SQLAlchemy
Flask-Migrate
python-decouple
